Sealed/internal
