+++ src/Microsoft.Diagnostics.Runtime/src/Common/HotColdRegions.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Diagnostics.Runtime
         /// </summary>
         public uint ColdSize { get; }
 
-        public HotColdRegions(ulong hotStart, uint hotSize, ulong coldStart, uint coldSize)
+        internal HotColdRegions(ulong hotStart, uint hotSize, ulong coldStart, uint coldSize)
         {
             HotStart = hotStart;
             HotSize = hotSize;
