+++ src/Microsoft.Diagnostics.Runtime/src/Extensions/ClrElementTypeExtensions.cs
