+++ src/Microsoft.Diagnostics.Runtime/src/Extensions/ClrHandleKindExtension.cs
