+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ClrDacType.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Diagnostics.Runtime.Builders;
 using Microsoft.Diagnostics.Runtime.DacInterface;
 using System;
 using System.Collections.Generic;
