+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/IClrThreadHelpers.cs
@@ -11,7 +11,7 @@ using System.Runtime.InteropServices;
 
 namespace Microsoft.Diagnostics.Runtime.Implementation
 {
-    public interface IClrThreadHelpers
+    internal interface IClrThreadHelpers
     {
         IDataReader DataReader { get; }
         IEnumerable<IClrStackRoot> EnumerateStackRoots(ClrThread thread);
