+++ src/Microsoft.Diagnostics.Runtime/src/Utilities/PEImage/PEImage.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Diagnostics.Runtime.Implementation;
 using System;
 using System.Buffers;
 using System.Collections.Generic;
